+++ src/Command/UpgradeCommand.php
@@ -2,13 +2,14 @@
 
 namespace OmekaCli\Command;
 
+use Zend_Registry;
+use Omeka_Db_Migration_Manager;
+use Omeka_Plugin_Installer;
 use OmekaCli\Application;
 use OmekaCli\UIUtils;
 
 class UpgradeCommand extends AbstractCommand
 {
-    protected $plugins;
-
     public function getDescription()
     {
         return 'upgrade Omeka';
@@ -33,183 +34,110 @@ class UpgradeCommand extends AbstractCommand
             return 1;
         }
 
-        if (!is_dir(BASE_DIR . '/.git')) {
-            $this->logger->Error('omeka-cli needs a git repo to upgrade Omeka');
+        if (!file_exists(BASE_DIR . '/.git')) {
+            $this->logger->error('omeka-cli needs a git repo to upgrade Omeka');
 
             return 1;
         }
 
         $this->logger->info('checking for updates');
-        $lastVersion = $this->getNewVersion();
-        if (!$lastVersion) {
-            $this->logger->error('Omeka is already up-to-date');
+        $latestVersion = $this->getLatestVersion();
+        if (version_compare(OMEKA_VERSION, $latestVersion) >= 0) {
+            $this->logger->notice('Omeka is already up-to-date');
 
-            return 1;
+            return 0;
         }
 
         $this->logger->info('saving database');
-        if ($this->saveDb()) {
+        if (false === $this->saveDb()) {
             $this->logger->error('database dumping failed');
 
             return 1;
         }
-        $this->logger->info('database saved in ~/.omeka-cli/backups/omeka_db_backup.sql');
-
-        $this->logger->info('deactivating plugins');
-        $this->deactivatePlugins();
-
-        $this->logger->info('saving Omeka');
-        if ($this->saveOmeka()) {
-            $this->logger->error('cannot save Omeka');
-
-            return 1;
-        }
-        $this->logger->info('Omeka saved in ~/.omeka-cli/backups/Omeka');
 
         $this->logger->info('upgrading Omeka');
-        if ($this->upgradeOmeka($lastVersion)) {
+        if (false === $this->upgradeOmeka($latestVersion)) {
             $this->logger->error('cannot upgrade Omeka');
-            $this->logger->info('recovering Omeka and its database');
-            if ($this->recover()) {
-                $this->logger->error('cannot recover Omeka or its database');
-            } else {
-                $this->logger->info('recovery successful');
-            }
 
             return 1;
         }
 
         $this->logger->info('upgrading database');
-        $this->upgradeDb($lastVersion);
+        if (false === $this->upgradeDb($latestVersion)) {
+            $this->logger->error('Failed to upgrade database');
+
+            return 1;
+        }
 
         $this->logger->info('upgrade successful');
 
         return 0;
     }
 
-    protected function getNewVersion()
+    protected function getLatestVersion()
     {
-        $lastVersion = shell_exec('git ls-remote --tags | grep -ho \'v[0-9]\+\(\.[0-9]\+\)*\' | cut -dv -f2 | tail -n1');
-        $lastVersion = trim($lastVersion);
+        $baseDir = BASE_DIR;
+        $latestTag = rtrim(`git ls-remote -q --tags --refs https://github.com/omeka/Omeka 'v*' | cut -f 2 | sed 's|refs/tags/||' | sort -rV | head -n1`);
+        $latestVersion = ltrim($latestTag, 'v');
 
-        return version_compare(OMEKA_VERSION, $lastVersion) < 0 ? $lastVersion : null;
+        return $latestVersion;
     }
 
     protected function saveDb()
     {
-        if (!is_dir(getenv('HOME') . '/.omeka-cli/backups')) {
-            if (!is_dir(getenv('HOME') . '/.omeka-cli')) {
-                mkdir(getenv('HOME') . '/.omeka-cli');
-            }
-            mkdir(getenv('HOME') . '/.omeka-cli/backups');
-        }
-        $lines = file(BASE_DIR . '/db.ini', FILE_IGNORE_NEW_LINES);
-        $lines = array_filter(
-            $lines,
-            function ($var) {
-                return preg_match('/(host|username|password|dbname)/', $var);
-            }
-        );
-        foreach ($lines as $line) {
-            $line = str_replace(' ', '', $line);
-            $line = str_replace('"', '', $line);
-            $line = explode('=', $line);
-            $infos[$line[0]] = $line[1];
+        $backupsDir = getenv('HOME') . '/.omeka-cli/backups';
+        if (!is_dir($backupsDir)) {
+            mkdir($backupsDir, 0777, true);
         }
 
-        exec('mysqldump -h\'' . $infos['host'] . '\''
-           . ' -u\'' . $infos['username'] . '\''
-           . ' -p\'' . $infos['password'] . '\''
-           . ' \'' . $infos['dbname'] . '\''
-           . ' > ~/.omeka-cli/backups/omeka_db_backup.sql', $out, $ans);
+        $db = parse_ini_file(BASE_DIR . '/db.ini');
+        $dest = sprintf('%s/sql/%s-%s.sql.gz', $backupsDir, $db['dbname'], date('YmdHis'));
+        if (!is_dir(dirname($dest))) {
+            mkdir(dirname($dest), 0777, true);
+        }
 
-        return $ans;
+        $mysqldump_cmd = 'mysqldump --host=' . escapeshellarg($db['host'])
+           . ' --user=' . escapeshellarg($db['username'])
+           . ' --password=' . escapeshellarg($db['password'])
+           . ' ' . escapeshellarg($db['dbname'])
+           . ' | gzip -c > ' . escapeshellarg($dest);
+
+        exec($mysqldump_cmd, $out, $ans);
+        if ($ans !== 0) {
+            $this->logger->error('mysqldump failed');
+
+            return false;
+        }
+
+        $this->logger->notice('database saved in {dest}', array('dest' => $dest));
+
+        return true;
     }
 
-    protected function deactivatePlugins()
+    protected function upgradeOmeka($version)
     {
-        $this->plugins = get_db()->getTable('Plugin')->findBy(array('active' => 1));
-        foreach ($this->plugins as $plugin) {
-            $broker = $plugin->getPluginBroker();
-            $loader = new \Omeka_Plugin_Loader($broker,
-                                               new \Omeka_Plugin_Ini(PLUGIN_DIR),
-                                               new \Omeka_Plugin_Mvc(PLUGIN_DIR),
-                                               PLUGIN_DIR);
-            $installer = new \Omeka_Plugin_Installer($broker, $loader);
-            $installer->deactivate($plugin);
-        }
+        $baseDir = escapeshellarg(BASE_DIR);
+        exec("git -C $baseDir stash save -q 'Stash made by omeka-cli upgrade'");
+        exec("git -C $baseDir fetch -q origin");
+        exec("git -C $baseDir reset --hard -q v$version", $out, $ans);
+
+        return $ans === 0;
     }
 
-    protected function saveOmeka()
+    protected function upgradeDb($version)
     {
-        if (!is_dir(getenv('HOME') . '/.omeka-cli/backups')) {
-            if (!is_dir(getenv('HOME') . '/.omeka-cli')) {
-                mkdir(getenv('HOME') . '/.omeka-cli');
+        $migrationMgr = Omeka_Db_Migration_Manager::getDefault();
+        try {
+            if ($migrationMgr->canUpgrade()) {
+                $migrationMgr->migrate();
+                set_option(Omeka_Db_Migration_Manager::VERSION_OPTION_NAME, $version);
             }
-            mkdir(getenv('HOME') . '/.omeka-cli/backups');
-        }
-        exec('cp -fr ' . BASE_DIR . ' ' . getenv('HOME') . '/.omeka-cli/backups/Omeka', $out, $ans);
+        } catch (\Exception $e) {
+            $this->logger->error('Database migration failed: {message}', array('message' => $e->getMessage()));
 
-        return $ans;
-    }
-
-    protected function upgradeOmeka($tag)
-    {
-        exec('git pull --rebase origin v' . $tag, $out, $ans);
-
-        return $ans;
-    }
-
-    protected function upgradeDb($ver)
-    {
-        $migrationMgr = new \Omeka_Db_Migration_Manager(get_db(), UPGRADE_DIR);
-        if ($migrationMgr->canUpgrade()) {
-            $migrationMgr->migrate();
-            set_option('omeka_version', $ver);
-        }
-    }
-
-    protected function recover()
-    {
-        if (!is_dir(getenv('HOME') . '/.omeka-cli/backups')) {
-            $this->logger->error('no backups directory found');
-
-            return 1;
-        }
-        if (count(scandir(getenv('HOME') . '/.omeka-cli/backups')) == 2) {
-            $this->logger->error('no backup found');
-
-            return 1;
-        }
-        $lines = file(BASE_DIR . '/db.ini', FILE_IGNORE_NEW_LINES);
-        $lines = array_filter(
-            $lines,
-            function ($var) {
-                return preg_match('/(host|username|password|dbname)/', $var);
-            }
-        );
-        foreach ($lines as $line) {
-            $line = str_replace(' ', '', $line);
-            $line = str_replace('"', '', $line);
-            $line = explode('=', $line);
-            $infos[$line[0]] = $line[1];
+            return false;
         }
 
-        if (UIUtils::confirmPrompt('Do you want to recover Omeka?')) {
-            exec('cp -fr ' . getenv('HOME') . '/.omeka-cli/backups/Omeka' . ' ' . BASE_DIR, $out, $ans1);
-            if (UIUtils::confirmPrompt('Do you want to recover the database?')) {
-                exec('mysql -h\'' . $infos['host'] . '\''
-                   . ' -u\'' . $infos['username'] . '\''
-                   . ' -p\'' . $infos['password'] . '\''
-                   . ' \'' . $infos['dbname'] . '\''
-                   . ' < ~/.omeka-cli/backups/omeka_db_backup.sql', $out, $ans2);
-            } else {
-                return 1;
-            }
-        } else {
-            return 1;
-        }
-
-        return $ans1 | $ans2;
+        return true;
     }
 }